+++ Demo/macOS/TiledDemoDelegate.swift
@@ -402,3 +402,4 @@ extension TiledDemoDelegate: TiledSceneCameraDelegate {
     @objc public func rightMouseDown(event: NSEvent) {}
 }
 
+